refactor: seperate Schedule comps into own fil
